Move `AssignUser` command
